+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 # Version `<dev>`
 
+### Enhancements
+* If funnel music is disabled, keep the base music playing while inside funnels.
+
 ### Bugfixes
 * Do not composite signage VTFs with PeTI backgrounds.
 * Fix signage not exporting if the UI was not opened at least once before exporting.